+++ src/Simplify.Web.Tests/Model/Binding/Parsers/ListToModelParserTestTypes/TestModelWithIdProperty.cs
@@ -1,0 +1,6 @@
+namespace Simplify.Web.Tests.Model.Binding.Parsers.ListToModelParserTestTypes;
+
+public class TestModelWithIdProperty
+{
+	public int Id { get; set; }
+}